services/notification_service/requirements.txt
kafka-python
minio
reportlab